requirements.txt
pytest
pylint
pytest-pylint
boto3
aws-cdk.aws-lambda
aws-cdk.aws_codecommit 
aws-cdk.aws-events
aws-cdk.aws-iam
aws-cdk.core
aws-cdk.aws_events_targets